integrate postgres

+++ Back-End/controllers/foodController.js
@@ -5,15 +5,14 @@ import fs from 'fs'
 const addFood = async (req, res) => {
   let image_filename = `${req.file.filename}`
 
-  const food = new foodModel({
-    name: req.body.name,
-    description: req.body.description,
-    price: req.body.price,
-    image: image_filename,
-    category: req.body.category,
-  })
   try {
-    await food.save()
+    await foodModel.create({
+      name: req.body.name,
+      description: req.body.description,
+      price: Number(req.body.price),
+      image: image_filename,
+      category: req.body.category,
+    })
     res
       .status(200)
       .json({ message: 'Food item added successfully', success: true })
@@ -26,7 +25,7 @@ const addFood = async (req, res) => {
 
 const listFood = async (req, res) => {
   try {
-    const foods = await foodModel.find({})
+    const foods = await foodModel.findAll()
     res.json({ success: true, data: foods })
   } catch (err) {
     res.status(500).json({ message: err.message })
@@ -38,13 +37,22 @@ const listFood = async (req, res) => {
 const removeFood = async (req, res) => {
   try {
     const food = await foodModel.findById(req.body.id)
-    fs.unlink(`uploads/${food.image}`, () => {} )
 
-    await foodModel.findByIdAndDelete(req.body.id)
-    res.status(200).json({ success: true, message: 'Food item deleted successfully' })
-    }catch (err) {
+    if (!food) {
+      return res
+        .status(404)
+        .json({ success: false, message: 'Food item not found' })
+    }
+
+    fs.unlink(`uploads/${food.image}`, () => {})
+    await foodModel.deleteById(req.body.id)
+    res
+      .status(200)
+      .json({ success: true, message: 'Food item deleted successfully' })
+  } catch (err) {
     res.status(500).json({ message: err.message })
-    }}
+  }
+}
 
 
 export { addFood, listFood, removeFood }
